Allow visible common entries to be hidden (issue154).
As we can now have non-integer local entries (for unused result parameters), we needed to be able to define the type of a local variable. To handle this a Local class was added, and a new 'child_contains_data' was added to detect whether a child (as opposed to the common entry) is visible.