Handle 'end-sequenceof' entries in all sequenceof types (issue209).
This allows a count sequenceof entries to have an 'end-sequenceof', and
causes the decode to fail if we there isn't an end-sequenceof entry.

This is a change in functionality; previously the decode would succeed
if an end-sequenceof was used, but the buffer ran out (this was bad
logically, as it means it's not possible to fail the decode of null
terminated strings in a data constrained environment.