Extract the python decoding code from the core objects into a separate bdec.decode module.
This is cleaner from a design point of view, but also fixes the
regression test failures introduced by the 're-validation' change
introduced in commit ef3a7fcd431f142208785fdb289c73c66b196b01. The
failures were due to the local name of child paramters changing
depending on which top level entry was doing the decoding; sometimes a
variable would be unneeded (and thus take on one name and local type),
and a be needed when decoded from another location (resulting in a
different parameter local name and type).