Fix encoding of common entries whose input parameters are unused in a given context.
Previously, when a common entry that had an ouput parameter which was
only used in some contexts, the encode from the context where it wasn't
used would be mocked using the parameter where it was used (instead of
null). This meant these mocked objects were encoded with either
uninitialised memory (as is the case in the 095 regression test), or a
value from another instance of itself.

We fix this by explicitly detecting the local names of all parameters
(not just output parameters), and treating all unknown parameters as
unused.