Consider the number of entries decoded when selecting the best guess.
A useful design is to have a packet type in a common section, then a
choice of the packet types with a sequence checking the correct packet
type. If the next field then fails to decode, it has effectively decoded
zero bits, despite correctly matching. This change makes it also
consider the number of entries successfully decoded when choosing the
best guess.