Fix memory read errors when an end-sequenceof is embedded inside multiple choices.
The inner choice doesn't always set the 'end-sequenceof', as it is only
set in one of the options. Previously the outer choice would copy the
inner 'end-sequenceof' to the outer 'end-sequenceof', but as it isn't
always initialised, it would sometimes be reading unitialised memory.

This is fixed by passing the original parameter into the child choice
when the type is the same; this is always true for 'end-seqenceof'. It
was only an error for 'end-sequenceof', as all other output parameter
types are required to be passed out by every option.