Correctly check the constraints when encoding fields.
We only want to encode the field if it matches the constraint; in the given
test case, we only want to encode the field when its value is greater than
forty (otherwise we want to encode using the second option).

While this may seem a little arbitrary, this is extracted from a real
specification (the AIS AIVDM specification).