Improve the error message when encoding fails.
Previously the best guess was chosen purely on how much data each option
had managed to encode. When encoding a choice, the option can start to
encode (based on the option name) before encoding any data, so we know
it's a better guess.

The code now selects primarily on the amount of data encoded, but falls
back to the number of entries succesfully queried on failure.